% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_dict_flat}
\alias{convert_dict_flat}
\title{Convert 'Quanteda' Dictionary to Flat Hierarchy Data Frame}
\usage{
convert_dict_flat(dict)
}
\arguments{
\item{dict}{\code{dictionary} from \code{quanteda}}
}
\value{
\code{data.frame}
}
\description{
Convert 'Quanteda' Dictionary to Flat Hierarchy Data Frame
}
