% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{convert_dict_hier}
\alias{convert_dict_hier}
\title{Convert 'Quanteda' Dictionary to 'd3.js' Hierarchy}
\usage{
convert_dict_hier(dict, idname = "name")
}
\arguments{
\item{dict}{\code{dictionary} from \code{quanteda}}

\item{idname}{\code{string} for the object property that contains the dictionary text; default is \code{"name"}}
}
\value{
\code{list} in `d3` hierarchy form
}
\description{
Convert 'Quanteda' Dictionary to 'd3.js' Hierarchy
}
